src/requirements.txt
aiohttp==3.9.1
aiosignal==1.3.1
async-timeout==4.0.3
attrs==23.1.0
flake8==6.1.0
frozenlist==1.4.0
idna==3.6
isort==5.12.0
loguru==0.7.2
mccabe==0.7.0
multidict==6.0.4
pycodestyle==2.11.1
pyflakes==3.1.0
python-dotenv==1.0.0
tqdm==4.66.1
yarl==1.9.3
